move all wrapped types out to module for deletion (#27)
